+++ tokens/usecase/pricing/worker/pricing_worker_test.go
@@ -2,7 +2,6 @@ package worker_test
 
 import (
 	"fmt"
-	"os"
 	"testing"
 	"time"
 
@@ -98,7 +97,7 @@ func (s *PricingWorkerTestSuite) TestUpdatePricesAsync() {
 			s.Require().NoError(err)
 
 			// Create a pricing worker
-			pricingWorker := worker.New(mainnetUsecase.Tokens, defaultQuoteDenom, &log.NoOpLogger{})
+			pricingWorker := worker.New(mainnetUsecase.Tokens, defaultQuoteDenom, defaultPricingConfig.WorkerMinPoolLiquidityCap, &log.NoOpLogger{})
 
 			// Create a mock listener
 			mockPricingUpdateListener := mocks.NewPricingListenerMock(time.Second * 5)
@@ -129,10 +128,10 @@ func (s *PricingWorkerTestSuite) TestUpdatePricesAsync() {
 }
 
 func (s *PricingWorkerTestSuite) TestGetPrices_Chain_FindUnsupportedTokens() {
-	env := os.Getenv("CI_SQS_PRICING_WORKER_TEST")
-	if env != "true" {
-		s.T().Skip("This test exists to identify which mainnet tokens are unsupported")
-	}
+	// env := os.Getenv("CI_SQS_PRICING_WORKER_TEST")
+	// if env != "true" {
+	// 	s.T().Skip("This test exists to identify which mainnet tokens are unsupported")
+	// }
 
 	viper.SetConfigFile("../../../../config.json")
 	err := viper.ReadInConfig()
@@ -151,7 +150,7 @@ func (s *PricingWorkerTestSuite) TestGetPrices_Chain_FindUnsupportedTokens() {
 	s.Require().NoError(err)
 
 	// Create a pricing worker
-	pricingWorker := worker.New(mainnetUsecase.Tokens, defaultQuoteDenom, &log.NoOpLogger{})
+	pricingWorker := worker.New(mainnetUsecase.Tokens, defaultQuoteDenom, config.Pricing.WorkerMinPoolLiquidityCap, &log.NoOpLogger{})
 
 	// Create a mock listener
 	mockPricingUpdateListener := mocks.NewPricingListenerMock(time.Minute * 5)